+++ release/changes.md
@@ -1,1 +1,1 @@
-- [NEW] TBD
+- [NEW] Include realized build time savings in experiment summary